% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpmn-widget.R
\name{bpmn_widget-shiny}
\alias{bpmn_widget-shiny}
\alias{bpmnWidgetOutput}
\alias{renderBpmnWidget}
\title{Shiny bindings for bpmn_widget}
\usage{
bpmnWidgetOutput(outputId, width = "100\%", height = "420px")

renderBpmnWidget(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output id.}

\item{width, height}{Widget size (CSS units).}

\item{expr}{Expression returning a [bpmn_widget()].}

\item{env, quoted}{Standard shiny render arguments.}
}
\description{
Shiny bindings for bpmn_widget
}
