% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbox.R
\name{inbox_dir}
\alias{inbox_dir}
\title{The inbox directory of a store}
\usage{
inbox_dir(store = ".runs")
}
\arguments{
\item{store}{Store directory.}
}
\value{
The path (not created).
}
\description{
The inbox directory of a store
}
