% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instance.R
\name{instance_definition}
\alias{instance_definition}
\title{The definition an instance was opened with}
\usage{
instance_definition(store = ".runs", instance = "instance")
}
\arguments{
\item{store}{Store directory.}

\item{instance}{Instance id to filter by, or `NULL` for all instances.}
}
\value{
The wide process table as stamped by [start_instance()], or `NULL`
  if the instance has not been opened.
}
\description{
The definition an instance was opened with
}
